requirements.txt
fastapi
uvicorn
pydantic
requests
playwright
beautifulsoup4
lxml
PyPDF2
python-pptx
pandas
openpyxl
python-docx
python-dotenv
langchain
langchain-huggingface
langchain-openai
langchain-community
openai
sentence-transformers
faiss-cpu